Changed datapackage.json to to conform to latest Table Schema spec.
The official specification for the Data Package schema from
[Frictionless data](http://specs.frictionlessdata.io/),
includes a spec for a [Table Schema](http://specs.frictionlessdata.io/table-schema)
which is being used by their official validator libs.

The validation libraries are using a JSON schema that enforces both
**primaryKey**,  **foreignKeys.fields** and **foreignKeys.reference.fields**
attributes to have values of type **array**.  According to the latest
Table Schema spec values can be either of type **string** or **array**,
but it is also stated that from this version if there is only one value,
it should be used as a single element array.