Validate some of the json examples
https://github.com/openreferral/specification/issues/355

Creates a github workflow for this, that's separate to the datapackage
examples.

Compiles the jsonschemas to include additionalProperties as false
everywhere, so we check we're not using any fields that aren't in the
schema.

Includes simple files that match a specific schema. Doesn't include
openapi responses, as the schema for these needs to be pulled out from
within openapi.json.