Docs: Tidied changelog headers
The HSDS Changelog headers attempt to link off to further information
surrounding the release of that version. However, versions 1.0 downwards
linked off to private google docs.

HSDS has only been semi-consistent in creating releases on Github for
versions. Some version headers link to issue queries filtered by
milestones. This is acceptable where no release exists, however others
link off to private, non-public, Google docs.

I have replaced the 1.0 link to the official GH release link, and
removed the others for which there is no milestone, label, or release
tag.

This means that for versions `0.x` down, there are currently no header
links. Links to the GDocs remain in the body of the changelog for that
version. In the future, if these details are pulled through onto a
public page or backported into Github; we can modify the headings with
new links.