Add .readthedocs.yaml, upgrade Python to 3.9 and upgrade package
Old version of python was 3.8 or lower, but this isn't supported any more.

To upgrade to 3.9 need to upgrade transifex-client,
as the old version of that we were using isn't supported in 3.9.

If you run pip-compile there are actually layout changes to requirements.txt
but I edited many of them back so we can see that other packages don't
change in this commit.

Need to use an older version of setuptools, or installing dependencies fails

https://github.com/openreferral/specification/issues/538