Tweak schema order so no 2 schemas have the same order
Previously url, unit and service_capacity all had order 99

This created a problem, as it was essentially random what order
the elements in files in schema/compiled would appear in.

Tweak order to 97,98 & 99 to make sure there is a set order.
Order choosen to match existing files in schema/compiled.