schema: OpenAPI Tag Definitions
Added the `tags` property to the OpenAPI definition to encapsulate and
define the tags which will be used to tag the endpoints.

Since the OpenAPI Tags are used to implement, rather than define, the
codelist; the `externalDocs.url` property points to the definition of
the codelist in the HSDS Docs site.