schema: updated simple/schema.csv
Fixes #559

simple/schema.csv is generated via the schemas-to-csv command in the
HSDS Schema tools script, however this is not automatically run as part
of a build process here.

This has resulted in the simple/schema.csv serialization drifting behind
the canonical schemas i.e. it is still representative of 3.0 and not 3.1
or any subsequent PATCHes.

This commit brings the schema.csv file in line with HSDS 3.1 by
auto-generating it from the current schemas.

Something to note: the schemas-to-csv command requires a file called
"fields_for_csv.csv", which is currently placed inside the HSDS Schema
Tools repo. When running the command inside this (the specification)
repo, one encounters an error.